+++ dhis-2/dhis-support/dhis-support-db-migration/src/main/resources/org/hisp/dhis/db/migration/2.42/V2_42_30__Create_uid_check_constraint.sql
@@ -1,9 +1,3 @@
-CREATE DOMAIN uid_type AS VARCHAR(11)
-    CHECK (
-        VALUE ~ '^[A-Za-z]' AND    -- First character must be a letter
-        VALUE ~ '^[A-Za-z0-9]{11}$' -- All 11 characters must be alphanumeric
-        );
-
 create or replace function check_uid(uid varchar) returns boolean
     language plpgsql
 as
@@ -12,6 +6,9 @@ begin
      return uid ~ '^[A-Za-z]' and uid ~ '^[A-Za-z0-9]{11}$';
 end;
 $$;
+
+CREATE DOMAIN uid_type AS VARCHAR(11)
+    CHECK (check_uid(VALUE));
 
 alter table potentialduplicate add constraint potentialduplicate_check_uid CHECK (check_uid(uid));
 alter table potentialduplicate add constraint potentialduplicate_check_original CHECK (check_uid(original));
